toolkit/mozapps/update/tests/chrome/test_0083_error_patchApplyFailure_partial_complete.xul

<?xml version="1.0"?>
<!--
/* Any copyright is dedicated to the Public Domain.
 * http://creativecommons.org/publicdomain/zero/1.0/
 */
-->

<?xml-stylesheet href="chrome://global/skin" type="text/css"?>
<?xml-stylesheet href="chrome://mochikit/content/tests/SimpleTest/test.css" type="text/css"?>

<window title="Update Wizard pages: error patching, download, and finished (partial failed and download complete)"
        xmlns="http://www.mozilla.org/keymaster/gatekeeper/there.is.only.xul"
        onload="runTestDefault();">
<script type="application/javascript"
        src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"/>
<script type="application/javascript"
        src="utils.js"/>

<script type="application/javascript">
<![CDATA[

const TESTS = [ {
  pageid: PAGEID_ERROR_PATCHING,
  buttonClick: "next"
}, {
  pageid: PAGEID_DOWNLOADING,
  extraStartFunction: createContinueFile
}, {
  pageid: PAGEID_FINISHED,
  buttonClick: "extra1",
  extraStartFunction: removeContinueFile
} ];

function runTest() {
  debugDump("entering");

  removeContinueFile();

  // Specify the url to update.sjs with a slowDownloadMar param so the ui can
  // load before the download completes.
  let slowDownloadURL = URL_HTTP_UPDATE_XML + "?slowDownloadMar=1";
  let patchProps = {type: "partial",
                    state: STATE_PENDING};
  let patches = getLocalPatchString(patchProps);
  patchProps = {url: slowDownloadURL,
                selected: "false"};
  patches += getLocalPatchString(patchProps);
  let updateProps = {isCompleteUpdate: "false"};
  let updates = getLocalUpdateString(updateProps, patches);

  writeUpdatesToXMLFile(getLocalUpdatesXMLString(updates), true);

  writeUpdatesToXMLFile(getLocalUpdatesXMLString(""), false);
  writeStatusFile(STATE_FAILED_CRC_ERROR);
  reloadUpdateManagerData();

  testPostUpdateProcessing();
}

]]>
</script>

<body xmlns="http://www.w3.org/1999/xhtml">
  <p id="display"></p>
  <div id="content" style="display: none"></div>
  <pre id="test"></pre>
</body>
</window>
